refactor : Price는 공통 사용되므로 common영역으로 이동
