+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -1,5 +1,7 @@
 package kitchenpos.product.domain;
 
+import kitchenpos.common.domain.Price;
+
 import javax.persistence.*;
 import java.math.BigDecimal;
